Tetris using Pyqt5
In Tetris, players complete lines by moving differently shaped pieces (tetrominoes), which descend onto the playing field. The completed lines disappear and grant the player points, and the player can proceed to fill the vacated spaces. The game ends when the playing field is filled. The longer the player can delay this inevitable outcome, the higher their score will be.
Tetris is primarily composed of a field of play in which pieces of different geometric forms, called "tetrominoes", descend from the top of the field.[1] During this descent, the player can move the pieces laterally and rotate them until they touch the bottom of the field or land on a piece that had been placed before it.[2] The player can neither slow down the falling pieces nor stop them. The objective of the game is to use the pieces to create as many vertical lines of blocks as possible. When a line is completed, it disappears, and the blocks placed above fall one rank. Completing lines grants points. You can use ctrl+P to pause and space