Typing Speed Test
This repository has a project of a typing speed tester.

To make this, you need the Tkinter library which will act as its Graphic User Interface (GUI)

In this, you will see a non-resizeable window containing a marquee, a timer, and a label displaying your score. You will also have a word displayed. You will have to type the exact word in the Entry box. If you type correctly, your score increases. If you did not type it correctly, the score will not change. If the timer finishes a minute, it will display your final score.

Enjoy!